provider: update add/remove proposal handling; mv legacy to separate file
